update format and mount

+++ utils/format_and_mount.sh
@@ -1,18 +1,28 @@
 # script to format mount and copy data.
 
+cwd=$(pwd)
+
+parent=$(pwd)/..
+
+if [[ ! -d "$parent/memory-pressure-benchmarks" ]]
+then
+	echo "Please rename current directory to memory-pressure-benchmarks"
+	exit 0
+fi
+
 # remove a leftover instance mount
-rm -rf memory-pressure-benchmarks-metal
+rm -rf $parent/memory-pressure-benchmarks-metal
 
 # format the mount
 sudo mkfs -t xfs /dev/nvme2n1
 
-mkdir memory-pressure-benchmarks-metal
+mkdir $parent/memory-pressure-benchmarks-metal
 # mount the nvme volumn
-sudo mount /dev/nvme2n1 memory-pressure-benchmarks-metal
+sudo mount /dev/nvme2n1 $parent/memory-pressure-benchmarks-metal
 # change ownsership of the volume
-sudo chown -R ubuntu memory-pressure-benchmarks-metal/
+sudo chown -R ubuntu $parent/memory-pressure-benchmarks-metal/
 
-git clone https://github.com/Tmonster/memory-pressure-benchmarks.git memory-pressure-benchmarks-metal
+git clone https://github.com/Tmonster/memory-pressure-benchmarks.git $parent/memory-pressure-benchmarks-metal
 
 # if you have an EBS volume, you can generate the data once, save it on the ebs volume, and transfer it
 # each time.
@@ -20,20 +30,20 @@ git clone https://github.com/Tmonster/memory-pressure-benchmarks.git memory-pres
 if [[ $# -gt 0 ]]
 then
 	echo "Creating data"
-	cd memory-pressure-benchmarks
+	cd $parent/memory-pressure-benchmarks
 	duckdb tpch-sf100.duckdb -c "install httpfs; load httpfs; call dbgen(sf=100);"
 	duckdb tpch-sf100.duckdb -c "export database 'tpch_data' (FORMAT CSV, HEADER 1);"
 	python3 hyper/load.py
-	cd ../memory-pressure-benchmarks-metal
-	cp ../memory-pressure-benchmarks/tpch-sf100.duckdb .
-	cp ../memory-pressure-benchmarks/tpch-sf100.hyper .
-elif [[ ! -d "memory-pressure-benchmarks-metal" ]]
+	cd $parent/memory-pressure-benchmarks-metal
+	cp $parent/memory-pressure-benchmarks/tpch-sf100.duckdb .
+	cp $parent/memory-pressure-benchmarks/tpch-sf100.hyper .
+elif [[ ! -d "$parent/memory-pressure-benchmarks-metal" ]]
 then
-	echo "no arguments passed. Copying data..."
-	echo "ERROR: directory ./memory-pressure-benchmarks-metal does not exist"
+	echo "ERROR: directory $parent/memory-pressure-benchmarks-metal does not exist"
 else
-	cd memory-pressure-benchmarks-metal/
-	cp ../memory-pressure-benchmarks/tpch-sf100.duckdb .
-	cp ../memory-pressure-benchmarks/tpch-sf100.hyper .
+	echo "no arguments passed. Copying data..."
+	cd $parent/memory-pressure-benchmarks-metal/
+	cp $parent/memory-pressure-benchmarks/tpch-sf100.duckdb .
+	cp $parent/memory-pressure-benchmarks/tpch-sf100.hyper .
 	cd ..
 fi
